+++ hlifealgo.h
@@ -173,7 +173,7 @@ public:
    virtual int isEmpty() ;
    virtual int hyperCapable() { return 1 ; }
    virtual void setMaxMemory(int m) ;
-   virtual int getMaxMemory() { return maxmem >> 20 ; }
+   virtual int getMaxMemory() { return (int)(maxmem >> 20) ; }
    virtual const char *setrule(const char *s) ;
    virtual const char *getrule() { return hliferules.getrule() ; }
    virtual void step() ;
